Route for candlestick chart is now dynamic, but no query string included yet. The other two charts only exist in my jupyter notebook.
