+++ Pipfile
@@ -16,6 +16,7 @@ numpy = "*"
 bokeh = "*"
 
 [dev-packages]
+requests = "*"
 
 [requires]
 python_version = "3.7"
