+++ stocks_api_project/tests/conftest.py
@@ -1,0 +1,1 @@
+# put all of your fixtures here if you want without needing to import
